+++ src/main/java/kitchenpos/domain/menu/MenuGroupRepository.java
@@ -1,11 +1,9 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.MenuGroup;
+package kitchenpos.domain.menu;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface MenuGroupDao {
+public interface MenuGroupRepository {
     MenuGroup save(MenuGroup entity);
 
     Optional<MenuGroup> findById(Long id);
