+++ src/main/java/kitchenpos/domain/menu/MenuRepository.java
@@ -1,11 +1,9 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.Menu;
+package kitchenpos.domain.menu;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface MenuDao {
+public interface MenuRepository {
     Menu save(Menu entity);
 
     Optional<Menu> findById(Long id);
