+++ src/main/java/kitchenpos/domain/order/OrderLineItemRepository.java
@@ -1,11 +1,9 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.OrderLineItem;
+package kitchenpos.domain.order;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface OrderLineItemDao {
+public interface OrderLineItemRepository {
     OrderLineItem save(OrderLineItem entity);
 
     Optional<OrderLineItem> findById(Long id);
