+++ src/main/java/kitchenpos/domain/order/OrderRepository.java
@@ -1,11 +1,9 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.Order;
+package kitchenpos.domain.order;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface OrderDao {
+public interface OrderRepository {
     Order save(Order entity);
 
     Optional<Order> findById(Long id);
