+++ src/main/java/kitchenpos/domain/table/TableGroup.java
@@ -1,7 +1,8 @@
-package kitchenpos.domain;
+package kitchenpos.domain.table;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import kitchenpos.domain.order.OrderTable;
 
 public class TableGroup {
     private Long id;
