+++ src/main/java/kitchenpos/infrastructure/repository/JdbcTemplateMenuProductRepository.java
@@ -1,6 +1,7 @@
-package kitchenpos.dao;
+package kitchenpos.infrastructure.repository;
 
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.menu.MenuProductRepository;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -16,14 +17,14 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public class JdbcTemplateProductDao implements ProductDao {
-    private static final String TABLE_NAME = "product";
-    private static final String KEY_COLUMN_NAME = "id";
+public class JdbcTemplateMenuProductRepository implements MenuProductRepository {
+    private static final String TABLE_NAME = "menu_product";
+    private static final String KEY_COLUMN_NAME = "seq";
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
-    public JdbcTemplateProductDao(final DataSource dataSource) {
+    public JdbcTemplateMenuProductRepository(final DataSource dataSource) {
         jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
         jdbcInsert = new SimpleJdbcInsert(dataSource)
                 .withTableName(TABLE_NAME)
@@ -32,14 +33,14 @@ public class JdbcTemplateProductDao implements ProductDao {
     }
 
     @Override
-    public Product save(final Product entity) {
+    public MenuProduct save(final MenuProduct entity) {
         final SqlParameterSource parameters = new BeanPropertySqlParameterSource(entity);
         final Number key = jdbcInsert.executeAndReturnKey(parameters);
         return select(key.longValue());
     }
 
     @Override
-    public Optional<Product> findById(final Long id) {
+    public Optional<MenuProduct> findById(final Long id) {
         try {
             return Optional.of(select(id));
         } catch (final EmptyResultDataAccessException e) {
@@ -48,23 +49,32 @@ public class JdbcTemplateProductDao implements ProductDao {
     }
 
     @Override
-    public List<Product> findAll() {
-        final String sql = "SELECT id, name, price FROM product";
+    public List<MenuProduct> findAll() {
+        final String sql = "SELECT seq, menu_id, product_id, quantity FROM menu_product";
         return jdbcTemplate.query(sql, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private Product select(final Long id) {
-        final String sql = "SELECT id, name, price FROM product WHERE id = (:id)";
+    @Override
+    public List<MenuProduct> findAllByMenuId(final Long menuId) {
+        final String sql = "SELECT seq, menu_id, product_id, quantity FROM menu_product WHERE menu_id = (:menuId)";
         final SqlParameterSource parameters = new MapSqlParameterSource()
-                .addValue("id", id);
+                .addValue("menuId", menuId);
+        return jdbcTemplate.query(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
+    }
+
+    private MenuProduct select(final Long id) {
+        final String sql = "SELECT seq, menu_id, product_id, quantity FROM menu_product WHERE seq = (:seq)";
+        final SqlParameterSource parameters = new MapSqlParameterSource()
+                .addValue("seq", id);
         return jdbcTemplate.queryForObject(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private Product toEntity(final ResultSet resultSet) throws SQLException {
-        final Product entity = new Product();
-        entity.setId(resultSet.getLong(KEY_COLUMN_NAME));
-        entity.setName(resultSet.getString("name"));
-        entity.setPrice(resultSet.getBigDecimal("price"));
+    private MenuProduct toEntity(final ResultSet resultSet) throws SQLException {
+        final MenuProduct entity = new MenuProduct();
+        entity.setSeq(resultSet.getLong(KEY_COLUMN_NAME));
+        entity.setMenuId(resultSet.getLong("menu_id"));
+        entity.setProductId(resultSet.getLong("product_id"));
+        entity.setQuantity(resultSet.getLong("quantity"));
         return entity;
     }
 }
